+++ test/hxtest_main.cpp
@@ -2,12 +2,10 @@
 // SPDX-License-Identifier: MIT
 // This file is licensed under the MIT license found in the LICENSE.md file.
 
-// Include everything first to catch conflicts.
 #include <hx/libhatchet.h>
 #include <hx/hxconsole.hpp>
 #include <hx/hxfile.hpp>
-#include <hx/hxtest.hpp> // May include Google Test.
-
+#include <hx/hxtest.hpp>
 #include "hxctest.h"
 
 HX_NS_USE
@@ -16,6 +14,7 @@ int test_main(int argc, char**argv);
 bool run_all_tests(const char* test_suite_filter);
 
 namespace {
+
 int hxs_test_assert_handler_count = 0;
 
 bool hxtest_assert_handler(void) {
@@ -27,37 +26,26 @@ bool hxtest_assert_handler(void) {
 TEST(hxtest_main, set_assert_handler) {
 	hxs_test_assert_handler_count = 0;
 	hxset_assert_handler(hxtest_assert_handler);
-
-	// volatile prevents the compiler from treating code after hxassert_always
-	// as unreachable.
 	const volatile bool failing = false;
 	hxassert_always((bool)failing, "hxtest_set_assert_handler_intentional");
-
 	hxset_assert_handler(hxnull);
 	EXPECT_EQ(hxs_test_assert_handler_count, 1);
 }
-
 bool run_all_tests(const char* test_suite_filter) {
 	hxlog_console("libhatchet 🪓🪓🪓 " LIBHATCHET_TAG "\n");
 	hxlog_console("C++: %d hardening: %d profiler: %d\n",
 		static_cast<int>(HX_CPLUSPLUS),
 		static_cast<int>(HX_HARDENING_MODE),
 		static_cast<int>(HX_USE_PROFILER));
-
-	// RUN_ALL_TESTS is a Google Test symbol. The optional filter arg is a
-	// libhatchet extension. Google Test implements filtering with weird hooks.
 #if !(HX_USE_GOOGLE_TEST)
-	const size_t tests_failing = static_cast<size_t>(RUN_ALL_TESTS(test_suite_filter));
+	const int tests_failing = RUN_ALL_TESTS(test_suite_filter);
 #else
-	const size_t tests_failing = static_cast<size_t>(RUN_ALL_TESTS()); (void)test_suite_filter;
+	const int tests_failing = RUN_ALL_TESTS(); (void)test_suite_filter;
 #endif
-
 #if HX_TEST_ERROR_HANDLING
 	const int hxs_expected_failures = 4;
-
 	hxassert_always(tests_failing == hxs_expected_failures,
 		"unexpected_failures Expected exactly %d tests to fail...", hxs_expected_failures);
-	// There are no asserts at level 3.
 	if(tests_failing == hxs_expected_failures) {
 		hxlog_handler(hxlog_level_warning,
 			"expected_failures Expected exactly %d tests to fail...", hxs_expected_failures);
@@ -67,24 +55,13 @@ bool run_all_tests(const char* test_suite_filter) {
 	return tests_failing == 0;
 #endif
 }
-
-// Ignored unless HX_USE_CONSOLE=1. Command line parameter to run all tests.
-// A test suite to filter by may be passed as an optional arg. E.g.
-// $ hxtest "runtests hxsort_test"
 hxconsole_command_named(run_all_tests, runtests);
-
 #if HX_USE_FILE_IO
-// Ignored unless HX_USE_CONSOLE=1. Command line parameter to execute stdin.
 hxconsole_command_named(*+[](void) -> bool { return hxconsole_exec_file(hxin); }, execstdin);
 #endif // HX_USE_FILE_IO
-
-// test_main - Command line console command dispatcher. Each parameter is
-// treated as a separate command.
 int test_main(int argc, char**argv) {
 	hxinit();
-
 	bool is_ok = true;
-// Console requires C++20.
 #if HX_USE_CONSOLE
 	if(argc > 1) {
 		for(int i=1; i<argc; ++i) {
@@ -98,22 +75,16 @@ int test_main(int argc, char**argv) {
 	(void)argc; (void)argv;
 	is_ok = run_all_tests(hxnull);
 #endif
-
-	// Logging and asserts are actually unaffected by a shutdown.
 	hxshutdown();
 	if(!is_ok) {
 		hxexit(EXIT_FAILURE);
 	}
 	return is_ok ? EXIT_SUCCESS : EXIT_FAILURE;
 }
-
-// main - calls test_main.
 int main(int argc, char**argv) {
 	testing::InitGoogleTest(&argc, argv);
-
 #if (HX_USE_GOOGLE_TEST) && ((HX_HARDENING_MODE) == HX_HARDENING_MODE_DEBUG)
 	GTEST_FLAG_SET(break_on_failure, true);
 #endif
-
 	return test_main(argc, argv);
 }
